+++ core/src/main/java/com/threerings/getdown/data/SysProps.java
@@ -101,6 +101,16 @@ public final class SysProps
         return Boolean.getBoolean("direct");
     }
 
+    /** If true, Getdown will always try to connect without proxy settings even it a proxy is set
+      * in {@code proxy.txt}. If direct access is possible it will not clear {@code proxy.txt}, it
+      * will preserve the settings. This is to support cases where a user uses a workstation in two
+      * different networks, one with proxy the other one without. They should not be asked for
+      * proxy settings again each time they switch back to the proxy network.
+      * Usage: {@code -Dtry_no_proxy}. */
+    public static boolean tryNoProxyFirst () {
+        return Boolean.getBoolean("try_no_proxy");
+    }
+
     /** Specifies the connection timeout (in seconds) to use when downloading control files from
       * the server. This is chiefly useful when you are running in versionless mode and want Getdown
       * to more quickly timeout its startup update check if the server with which it is